cf_get_articles_id/requirements.txt
numpy
watchdog==4.*
functions-framework==3.*
implicit==0.4.8
google-cloud-storage
google-cloud-bigquery
pandas
db-dtypes
